switch to using csvs
